I swapped out e.getID() with e.getKeyChar().
Some folks get a little uncomfortable at the idea, but I found that it is the most user-friendly way to accomplish it. You tell the user that, to control this thing, you must use "+" or "-". Then, simply let the user enter that in. Works better, especially for international users with non-english/american/uk keyboards.